Promoting several Periodic tests to Main suite
